tlda_requirements.txt
absl-py @ file:///C:/ci/absl-py_1624299085608/work
cached-property==1.5.2
carla==0.9.6
certifi==2021.5.30
charset-normalizer==2.0.4
cloudpickle==1.6.0
Cython==0.29.24
dm-env==1.5
dm-tree==0.1.6
einops==0.3.2
future==0.18.2
glfw==1.12.0
GPUtil==1.4.0
gym==0.19.0
h5py==3.4.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
kornia==0.5.10
labmaze==1.0.5
lxml==4.6.3
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.2
protobuf==3.17.3
PyOpenGL==3.1.5
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
requests==2.26.0
scipy==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
termcolor==1.1.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchgeometry==0.1.2
torchvision==0.10.0+cu111
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
xmltodict==0.12.0
